requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.2
docopt==0.6.2
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.10.0
incremental==21.3.0
kombu==4.6.11
msgpack==1.0.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.2.0
requests==2.25.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.1
zope.interface==5.3.0
